+++ CairoMakie/src/primitives.jl
@@ -318,18 +318,23 @@ function draw_atomic(scene::Scene, screen::Screen, @nospecialize(primitive::Scat
     ctx = screen.context
     positions = primitive[1][]
     isempty(positions) && return
-    size_model = transform_marker ? model : Mat4d(I)
+    markerspace = primitive.markerspace[]::Symbol
+    space = primitive.space[]::Symbol
+
+    size_model = transform_marker::Bool ? model::Mat4d : Mat4d(I)
+    if !isnothing(scene.float32convert) && Makie.is_data_space(markerspace)
+        size_model = Makie.scalematrix(scene.float32convert.scaling[].scale::Vec3d) * size_model
+    end
 
     font = to_font(to_value(get(primitive, :font, Makie.defaultfont())))
     colors = to_color(primitive.calculated_colors[])
-    markerspace = primitive.markerspace[]
-    space = primitive.space[]
+
     transfunc = Makie.transform_func(primitive)
     billboard = primitive.rotation[] isa Billboard
 
     return draw_atomic_scatter(scene, ctx, transfunc, colors, markersize, strokecolor, strokewidth, marker,
-                               marker_offset, rotation, model, positions, size_model, font, markerspace,
-                               space, clip_planes, billboard)
+                               marker_offset, rotation, model::Mat4d, positions, size_model, font, markerspace,
+                               space, clip_planes::Vector{Plane3f}, billboard)
 end
 
 function is_degenerate(M::Mat2f)
@@ -346,8 +351,8 @@ is_approx_zero(v::VecTypes) = any(x -> isapprox(x, 0), v)
 
 function draw_atomic_scatter(
         scene, ctx, transfunc, colors, markersize, strokecolor, strokewidth,
-        marker, marker_offset, rotation, model, positions, size_model, font,
-        markerspace, space, clip_planes, billboard
+        marker, marker_offset, rotation, model::Mat4d, positions, size_model::Mat4d, font,
+        markerspace::Symbol, space::Symbol, clip_planes::Vector{Plane3f}, billboard::Bool
     )
 
     transformed = apply_transform(transfunc, positions, space)
@@ -356,7 +361,7 @@ function draw_atomic_scatter(
         Makie.space_to_clip(scene.camera, space) *
         Makie.f32_convert_matrix(scene.float32convert, space) *
         model
-    model33 = size_model[Vec(1,2,3), Vec(1,2,3)]
+    model33 = size_model[Vec(1,2,3), Vec(1,2,3)]::Mat3d
 
     Makie.broadcast_foreach_index(view(transformed, indices), indices, colors, markersize, strokecolor,
             strokewidth, marker, marker_offset, remove_billboard(rotation)) do pos, col,
@@ -551,8 +556,9 @@ function draw_atomic(scene::Scene, screen::Screen, @nospecialize(primitive::Text
 
     draw_glyph_collection(
         scene, ctx, position, glyph_collection, remove_billboard(rotation),
-        model, space, markerspace, offset, primitive.transformation, transform_marker,
-        clip_planes
+        model::Mat4d, space::Symbol, markerspace::Symbol, offset,
+        primitive.transformation::Makie.Transformation,
+        transform_marker, clip_planes::Vector{Plane3f}
     )
 
     nothing
@@ -560,7 +566,7 @@ end
 
 function draw_glyph_collection(
         scene, ctx, positions, glyph_collections::AbstractArray, rotation,
-        model::Mat, space, markerspace, offset, transformation, transform_marker,
+        model, space, markerspace, offset, transformation, transform_marker,
         clip_planes
     )
 
@@ -590,7 +596,9 @@ function draw_glyph_collection(
 
     model = _deref(_model)
     model33 = transform_marker ? model[Vec(1, 2, 3), Vec(1, 2, 3)] : Mat3d(I)
-    id = Mat4f(I)
+    if !isnothing(scene.float32convert) && Makie.is_data_space(markerspace)
+        model33 = Makie.scalematrix(scene.float32convert.scaling[].scale::Vec3d)[Vec(1,2,3), Vec(1,2,3)] * model33
+    end
 
     glyph_pos = let
         # TODO: f32convert may run into issues here if markerspace is :data or
@@ -633,7 +641,7 @@ function draw_glyph_collection(
         end
 
         scale2 = scale isa Number ? Vec2d(scale, scale) : scale
-        glyphpos, mat, _ = project_marker(scene, markerspace, gp3, scale2, rotation, model33, id)
+        glyphpos, mat, _ = project_marker(scene, markerspace, gp3, scale2, rotation, model33)
 
         Cairo.save(ctx)
         set_font_matrix(ctx, mat)